requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
frozenlist==1.4.1
gevent==23.9.1
greenlet==3.0.0rc3
grequests==0.7.0
idna==3.6
matplotlib<3.10
multidict==6.0.5
mysql-connector-python==9.4.0
numpy==1.26.4
pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.0
yarl==1.9.4
zope.event==5.0
zope.interface==6.1
